refactor: streamline demo video handling and remove preloading functionality
- Moved demo video URL fetching to the IntroScreen for better user experience
- Removed preloading logic from RootLayout to prevent errors with expo-av
- Simplified useDemoVideo hook by eliminating unnecessary state management
- Updated video component to enhance loading behavior and UI consistency
- Deleted unused icon components related to video controls

These changes improve the overall structure and reliability of the demo video feature while aligning with the latest Expo and React best practices.
