+++ apps/expo/src/app/_layout.tsx
@@ -307,39 +307,6 @@ function RootLayoutContent() {
     routingInstrumentation.registerNavigationContainer(ref);
   }, [ref]);
 
-  // Preload demo video after a delay to avoid impacting startup
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      void (async () => {
-        try {
-          if (__DEV__) {
-            console.log("Starting demo video preload...");
-          }
-
-          const { preloadDemoVideo } = await import("~/services/videoPreload");
-          const { api } = await import(
-            "@soonlist/backend/convex/_generated/api"
-          );
-
-          // Fetch video URL from Convex and preload
-          const videoUrl = await convex.query(api.appConfig.getDemoVideoUrl);
-          if (videoUrl) {
-            if (__DEV__) {
-              console.log("Demo video URL fetched:", videoUrl);
-            }
-            preloadDemoVideo(videoUrl);
-            if (__DEV__) {
-              console.log("Demo video preload completed");
-            }
-          }
-        } catch (error) {
-          console.log("Failed to preload demo video:", error);
-        }
-      })();
-    }, 5000); // 5 second delay
-
-    return () => clearTimeout(timer);
-  }, []);
 
   // The share extension logic now specifically leads to /new
   useIntentHandler();
