--- apps/expo/src/components/icons/Pause.tsx
@@ -1,14 +1,0 @@
-import createLucideIcon from "./createLucideIcon";
-
-const Pause = createLucideIcon("Pause", [
-  [
-    "rect",
-    { x: "14", y: "4", width: "4", height: "16", rx: "1", key: "zuxfzm" },
-  ],
-  [
-    "rect",
-    { x: "6", y: "4", width: "4", height: "16", rx: "1", key: "1okwgv" },
-  ],
-]);
-
-export { Pause as default };
