--- apps/expo/src/components/icons/Play.tsx
@@ -1,7 +1,0 @@
-import createLucideIcon from "./createLucideIcon";
-
-const Play = createLucideIcon("Play", [
-  ["polygon", { points: "6 3 20 12 6 21 6 3", key: "1oa8hb" }],
-]);
-
-export { Play as default };
